{{cookiecutter.directory_name}}/users/models.py
from django.db import models
from django.contrib.auth.models import AbstractUser
from django.conf import settings

from PIL import Image

class User(AbstractUser):
    pass

class Account(models.Model):
    # settings.AUTH_USER_MODEL
    user = models.OneToOneField(settings.AUTH_USER_MODEL, on_delete=models.CASCADE)
    image = models.ImageField(default='default.jpg', upload_to='profile_photos')

    def __str__(self):
        return f'{self.user.username} Profile'

    def save(self, *args, **kwargs):
        super().save(*args, **kwargs)

        img = Image.open(self.image.path)

        if img.height > 300 or img.width > 300:
            output_size = (300, 300)
            img.thumbnail(output_size)
            img.save(self.image.path)
