{{cookiecutter.directory_name}}/requirements.txt
asgiref==3.4.1
crispy-bootstrap5==0.6
Django==3.2.9
django-crispy-forms==1.13.0
Pillow==8.4.0
pytz==2021.3
sqlparse==0.4.2